chore: Run static checks on CI
openactive-test-suite's npm test has been repurposed, so that this runs the static checks. The Booking API tests are now run with npm start. This fits with the NPM pattern where `npm test` tests the code itself and `npm start` runs the code. In this case, the code is tests, so the language can be confusing